refactor: split payment page into server and client component
